pyproject.toml
[project]
name = "real_estate_ai"
version = "0.1.0"
description = "AI agent for real estate property inquiries in Egypt"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-supervisor",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "langmem>=0.0.27",
    "langgraph-cli[inmem]>=0.2.0",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0"
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120